@@ -1,15 +1,9 @@
-module github.com/XiaoMiku01/douyin-live-go
+module github.com/zboyco/douyin-live-go
 
-go 1.20
+go 1.21
 
 require (
-	github.com/golang/protobuf v1.5.2
 	github.com/gorilla/websocket v1.5.0
-	github.com/tidwall/gjson v1.14.4
-	google.golang.org/protobuf v1.28.1
-)
-
-require (
-	github.com/tidwall/match v1.1.1 // indirect
-	github.com/tidwall/pretty v1.2.0 // indirect
+	github.com/pkg/errors v0.9.1
+	google.golang.org/protobuf v1.31.0
 )
